feat(web): fix translation
(cherry picked from commit 79a4411135e735e78237104588db041d825754d5)
(cherry picked from commit 92fe37ba94bee36eb56c407c68f5b0c99191a71f)
(cherry picked from commit bd78428316b83484c1774975e7e5e75bb6e3837f)
(cherry picked from commit 8c71f590a71fb6da19f517774cd19c3742d4e8ff)
(cherry picked from commit dc1b19bfe4170283cf34d620da212489e09c66d8)
(cherry picked from commit 4d74a3788a399321f62de8cf53abd87b40b33413)
(cherry picked from commit 285af99a0fd85d650f2a9eda11914e58d9218e7d)
(cherry picked from commit 2bd0869aa4c3e48c65eec143967cfce973f43a62)
