+++ web/public/locales/en/translation.json
@@ -54,7 +54,7 @@
   "Edit": "Edit ",
   "Generate": "generate",
   "InputTip": "please enter",
-  "IsRequired": "is required",
+  "IsRequired": " is required",
   "LimitSelect": "Please select at least one",
   "Logout": "Logout",
   "Name": "name",
